server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
numpy
PyPDF2
weaviate-client
python-multipart
nltk
